+++ Solar/core/scene/scene_camera.cpp
@@ -25,8 +25,8 @@ void SceneCamera::SetPerspective(float vertical_fov, float near_clip,
   RecalculateProjection();
 }
 
-void SceneCamera::SetViewportSize(uint32_t width, uint32_t height) {
-  aspect_ratio_ = static_cast<float>(width) / static_cast<float>(height);
+void SceneCamera::SetViewportSize(float width, float height) {
+  aspect_ratio_ = width / height;
   RecalculateProjection();
 }
