+++ Solar/core/scene/scene_camera.h
@@ -27,8 +27,7 @@ class SceneCamera : public Camera {
   void SetOrthoGraphic(float size, float near_clip, float far_clip);
   void SetPerspective(float vertical_fov, float near_clip, float far_clip);
 
-  // TODO(sylar): set to float
-  void SetViewportSize(uint32_t width, uint32_t height);
+  void SetViewportSize(float width, float height);
 
   float GetPerspectiveVerticalFOV() const;
   void SetPerspectiveVerticalFOV(float vertical_fov);
